django-backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-decouple==3.8
requests==2.32.4
django-cors-headers==4.8.0
pymongo==4.13.2
djongo==1.2.31
gunicorn==23.0.0
pandas==2.3.3
openpyxl==3.1.5
bcrypt==5.0.0
redis==5.0.1